Raise mimumum Faraday version to 0.15
There was a bug in our Appraisials stating 0.14.x passed - it was
actually testing with 0.15.x (due to the pessimistic operator missing the
last digit for the patch version).

Unfortunately the gem has never been compatible with 0.14.x, so
we're raising the version to the true minimum supported version - 0.15.x.

Resolves https://github.com/remove-bg/ruby/issues/3

https://guides.rubygems.org/patterns/#pessimistic-version-constraint